moved Pagination logic into MovieList and Slice
