+++ src/common/Pagination/index.js
@@ -1,13 +1,4 @@
-import { useEffect } from 'react';
-import { useSelector, useDispatch } from 'react-redux';
-import { useLocation, useHistory } from 'react-router-dom';
-import {
-	incrementPage,
-	decrementPage,
-	goToFirstPage,
-	pageNumberFromURL,
-	selectPageState,
-} from './paginationSlice';
+import { useDispatch } from 'react-redux';
 import {
 	Wrapper,
 	StyledButton,
@@ -21,26 +12,8 @@ import {
 	BoldText,
 } from './styled';
 
-const Pagination = () => {
+const Pagination = ({ currentPage, goToFirstPage, decrementPage, incrementPage }) => {
 	const dispatch = useDispatch();
-	const currentPage = useSelector(selectPageState);
-	const history = useHistory();
-	const location = useLocation();
-	const searchParams = new URLSearchParams(location.search);
-	const query = searchParams.get('page');
-
-	useEffect(() => {
-		if (query < 1) {
-			searchParams.set('page', 1);
-		} else {
-			dispatch(pageNumberFromURL(query));
-		}
-	}, [query]);
-
-	useEffect(() => {
-
-		history.push(`${location.pathname}?page=${currentPage}`);
-	}, [currentPage]);
 
 	return (
 		<Wrapper>
