+++ src/features/movies/MovieList/index.js
@@ -1,13 +1,42 @@
-import { useSelector } from "react-redux";
-import { Link } from "react-router-dom";
-import { selectMovieList, selectLoading } from "./movieListSlice";
+import { useEffect } from 'react';
+import { useSelector, useDispatch } from 'react-redux';
+import { useLocation, useHistory, Link } from "react-router-dom";
+import { 
+    incrementPage,
+    decrementPage,
+    goToFirstPage,
+    pageNumberFromURL,
+    selectMovieList, 
+    selectLoading,
+    selectPageState 
+} from "./movieListSlice";
 import { Section, SectionTitle } from "../../../common/Section/Section";
-import Pagination from '../../../common/Pagination/index';
 import { SectionWrapper, Tile } from "./styled";
+import Pagination from '../../../common/Pagination/index';
 
 function MovieList() {
+    const dispatch = useDispatch();
+    const currentPage = useSelector(selectPageState);
     const popularMovies = useSelector(selectMovieList);
     const loading = useSelector(selectLoading);
+
+    const history = useHistory();
+    const location = useLocation();
+    const searchParams = new URLSearchParams(location.search);
+    const query = searchParams.get('page');
+  
+    useEffect(() => {
+      if (query < 1) {
+        searchParams.set('page', 1);
+      } else {
+        dispatch(pageNumberFromURL(query));
+      }
+    }, [query]);
+  
+    useEffect(() => {
+  
+      history.push(`${location.pathname}?page=${currentPage}`);
+    }, [currentPage]);
 
     if (loading) {
         return <p>Loading...</p>;
@@ -39,8 +68,7 @@ function MovieList() {
                     ))}
                 </SectionWrapper>
             </Section>
-            {/* <Pagination /> */}
-        </>
+            <Pagination currentPage={currentPage} goToFirstPage={goToFirstPage} incrementPage={incrementPage} decrementPage={decrementPage}/>        </>
     );
 };
